Factor out list tests (#267)
* Factor out list tests

* Refactor to function
